requirements.txt
allure-pytest==2.13.0
allure-python-commons==2.13.0
annotated-types==0.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
dnspython==2.3.0
email-validator==2.1.0.post1
exceptiongroup==1.1.0
Faker==17.6.0
greenlet==3.0.3
idna==3.4
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pydantic==2.6.2
pydantic_core==2.16.3
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.24
tomli==2.0.1
typing_extensions==4.9.0
urllib3==1.26.14
